requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
attrs==23.2.0
black==24.2.0
cfgv==3.4.0
click==8.1.7
coverage==7.4.1
distlib==0.3.8
Django==5.0.2
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
factory-boy==3.3.0
Faker==23.2.1
filelock==3.13.1
flake8==7.0.0
identify==2.5.35
inflection==0.5.1
iniconfig==2.0.0
isort==5.13.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
pluggy==1.4.0
pre-commit==3.6.2
pycodestyle==2.11.1
pyflakes==3.2.0
PyJWT==2.8.0
pytest==8.0.1
pytest-cov==4.1.0
pytest-django==4.8.0
pytest-factoryboy==2.6.0
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
rpds-py==0.18.0
setuptools==69.1.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
uritemplate==4.1.1
virtualenv==20.25.0
